Move affine-transform to utils
