+++ cocos2d/core/CCNode.js
@@ -29,7 +29,7 @@ const BaseNode = require('./utils/base-node');
 const PrefabHelper = require('./utils/prefab-helper');
 const mathPools = require('./utils/math-pools');
 const renderEngine = require('./renderer/render-engine');
-const AffineTrans = require('./value-types/affine-transform');
+const AffineTrans = require('./utils/affine-transform');
 const math = renderEngine.math;
 const eventManager = require('./event-manager');
 const macro = require('./platform/CCMacro');
