+++ cocos2d/core/utils/affine-transform.js
